+++ satella/instrumentation/metrics/metric_types/__init__.py
@@ -1,4 +1,5 @@
-from .base import Metric, MetricLevel, LeafMetric, EmbeddedSubmetrics
+from .base import Metric, MetricLevel, LeafMetric, EmbeddedSubmetrics, DISABLED, RUNTIME, \
+    DEBUG, INHERIT
 from .cps import ClicksPerTimeUnitMetric
 from .simple import IntegerMetric, FloatMetric
 from .counter import CounterMetric
@@ -12,4 +13,5 @@ from .registry import register_metric, METRIC_NAMES_TO_CLASSES
 __all__ = ['Metric', 'LeafMetric', 'EmbeddedSubmetrics', 'ClicksPerTimeUnitMetric',
            'IntegerMetric', 'FloatMetric',
            'QuantileMetric', 'register_metric', 'METRIC_NAMES_TO_CLASSES', 'SummaryMetric',
-           'HistogramMetric', 'EmptyMetric', 'LinkfailMetric', 'CallableMetric', 'MetricLevel']
+           'HistogramMetric', 'EmptyMetric', 'LinkfailMetric', 'CallableMetric', 'MetricLevel',
+           'INHERIT', 'DEBUG', 'RUNTIME', 'DISABLED']